+++ CHANGELOG.md
@@ -7,6 +7,9 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 
 ## Unreleased
 
+### Fixed
+- Fix JSONPath iteration and referencing, introduced in release 2.1.0
+
 ## [2.1.1] - 2025-08-27
 
 ### Fixed